refactor moduls calls
