+++ js/components/map-filters.js
@@ -1,13 +1,13 @@
-import {activate, deactivate} from '../common/switch-form-activity.js';
+import * as Forms from '../common/forms.js';
 
 const mapFilters = document.querySelector('.map__filters');
 
-const activateMapFilters = () => {
-  activate(mapFilters, 'ad-form--disabled');
+const activate = () => {
+  Forms.activate(mapFilters, 'ad-form--disabled');
 };
 
-const deactivateMapFilters = () => {
-  deactivate(mapFilters, 'map__filters--disabled');
+const deactivate = () => {
+  Forms.deactivate(mapFilters, 'map__filters--disabled');
 };
 
-export {activateMapFilters, deactivateMapFilters};
+export {activate, deactivate};
